go.mod: rename module to full source
Typical go modules should have a module name that points to the source
of the code.
This is not required for compilation, but this is a convention that is
adopted for most projects.

The module name will also be used by the GNUmakefile of the project to
install the plugin later on for the dev target, so we change it in this
commit.